+++ sgp30/sgp30.c
@@ -33,182 +33,85 @@
 #include "sensirion_arch_config.h"
 #include "sensirion_common.h"
 #include "sensirion_i2c.h"
-#include "sgp_featureset.h"
 #include "sgp_git_version.h"
-
-#define SGP30_MAX_BUFFER_WORDS 3
-#define SGP30_VALID_IAQ_BASELINE(b) ((b) != 0)
 
 static const uint8_t SGP30_I2C_ADDRESS = 0x58;
 
 /* command and constants for reading the serial ID */
+#define SGP30_CMD_GET_SERIAL_ID 0x3682
 #define SGP30_CMD_GET_SERIAL_ID_DURATION_US 500
 #define SGP30_CMD_GET_SERIAL_ID_WORDS 3
-#define SGP30_CMD_GET_SERIAL_ID 0x3682
 
 /* command and constants for reading the featureset version */
+#define SGP30_CMD_GET_FEATURESET 0x202f
 #define SGP30_CMD_GET_FEATURESET_DURATION_US 1000
 #define SGP30_CMD_GET_FEATURESET_WORDS 1
-#define SGP30_CMD_GET_FEATURESET 0x202f
 
 /* command and constants for on-chip self-test */
+#define SGP30_CMD_MEASURE_TEST 0x2032
 #define SGP30_CMD_MEASURE_TEST_DURATION_US 220000
 #define SGP30_CMD_MEASURE_TEST_WORDS 1
 #define SGP30_CMD_MEASURE_TEST_OK 0xd400
-#define SGP30_CMD_MEASURE_TEST 0x2032
 
-static const struct sgp_otp_featureset sgp30_features_unknown = {
-    .profiles = NULL,
-    .number_of_profiles = 0,
-};
+/* command and constants for IAQ init */
+#define SGP30_CMD_IAQ_INIT 0x2003
+#define SGP30_CMD_IAQ_INIT_DURATION_US 10000
 
-enum sgp30_state_code { WAIT_STATE, MEASURING_PROFILE_STATE };
+/* command and constants for IAQ measure */
+#define SGP30_CMD_IAQ_MEASURE 0x2008
+#define SGP30_CMD_IAQ_MEASURE_DURATION_US 50000
+#define SGP30_CMD_IAQ_MEASURE_WORDS 2
 
-struct sgp30_info {
-    uint64_t serial_id;
-    uint16_t feature_set_version;
-};
+/* command and constants for getting IAQ baseline */
+#define SGP30_CMD_GET_IAQ_BASELINE 0x2015
+#define SGP30_CMD_GET_IAQ_BASELINE_DURATION_US 10000
+#define SGP30_CMD_GET_IAQ_BASELINE_WORDS 2
 
-static struct sgp30_data {
-    enum sgp30_state_code current_state;
-    struct sgp30_info info;
-    const struct sgp_otp_featureset *otp_features;
-    union {
-        uint16_t words[SGP30_MAX_BUFFER_WORDS];
-        uint64_t u64_value;
-    } buffer;
-} client_data;
+/* command and constants for setting IAQ baseline */
+#define SGP30_CMD_SET_IAQ_BASELINE 0x201e
+#define SGP30_CMD_SET_IAQ_BASELINE_DURATION_US 10000
 
-/**
- * unpack_signals() - unpack signals which are stored in
- *                    client_data.buffer.words
- * @profile:    The profile
- */
-static void unpack_signals(const struct sgp_profile *profile) {
-    int16_t i, j;
-    uint16_t data_words = profile->num_words;
-    uint16_t word_buf[SGP30_MAX_BUFFER_WORDS];
+/* command and constants for raw measure */
+#define SGP30_CMD_RAW_MEASURE 0x2050
+#define SGP30_CMD_RAW_MEASURE_DURATION_US 25000
+#define SGP30_CMD_RAW_MEASURE_WORDS 2
 
-    /* copy buffer */
-    for (i = 0; i < data_words; i++)
-        word_buf[i] = client_data.buffer.words[i];
+/* command and constants for setting absolute humidity */
+#define SGP30_CMD_SET_ABSOLUTE_HUMIDITY 0x2061
+#define SGP30_CMD_SET_ABSOLUTE_HUMIDITY_DURATION_US 10000
 
-    /* signals are in reverse order in the data buffer */
-    for (i = profile->num_words - 1, j = 0; i >= 0; i -= 1, j += 1)
-        client_data.buffer.words[j] = word_buf[i];
-}
+/* command and constants for getting TVOC inceptive baseline */
+#define SGP30_CMD_GET_TVOC_INCEPTIVE_BASELINE 0x20b3
+#define SGP30_CMD_GET_TVOC_INCEPTIVE_BASELINE_DURATION_US 10000
+#define SGP30_CMD_GET_TVOC_INCEPTIVE_BASELINE_WORDS 1
+
+/* command and constants for setting TVOC baseline */
+#define SGP30_CMD_SET_TVOC_BASELINE 0x2077
+#define SGP30_CMD_SET_TVOC_BASELINE_DURATION_US 10000
 
 /**
- * read_measurement() - reads the result of a profile measurement
+ * sgp30_check_featureset() - Check if the connected sensor has a certain FS
  *
- * Return:  STATUS_OK on success, an error code otherwise
+ * @needed_fs: The featureset that is required
+ *
+ * Return: STATUS_OK if the sensor has the required FS,
+ *         SGP30_ERR_INSUFFICIENT_FEATURE_SET if the sensor does not
+ *                                            have the required FS,
+ *         an error code otherwise
  */
-static int16_t read_measurement(const struct sgp_profile *profile) {
+static int16_t sgp30_check_featureset(uint16_t needed_fs) {
     int16_t ret;
+    uint16_t fs_version;
+    uint8_t product_type;
 
-    switch (client_data.current_state) {
-
-        case MEASURING_PROFILE_STATE:
-            ret = sensirion_i2c_read_words(SGP30_I2C_ADDRESS,
-                                           client_data.buffer.words,
-                                           profile->num_words);
-
-            if (ret)
-                /* Measurement in progress */
-                return STATUS_FAIL;
-
-            unpack_signals(profile);
-            client_data.current_state = WAIT_STATE;
-
-            return STATUS_OK;
-
-        default:
-            /* No command issued */
-            return STATUS_FAIL;
-    }
-}
-
-/**
- * sgp30_get_profile_by_number() - get a profile by its identifier number
- * @number      The number that identifies the profile
- *
- * Return:      A pointer to the profile or NULL if the profile does not exists
- */
-static const struct sgp_profile *sgp30_get_profile_by_number(uint16_t number) {
-    uint8_t i;
-    const struct sgp_profile *profile = NULL;
-
-    for (i = 0; i < client_data.otp_features->number_of_profiles; i++) {
-        profile = client_data.otp_features->profiles[i];
-        if (number == profile->number)
-            break;
-    }
-
-    if (i == client_data.otp_features->number_of_profiles) {
-        return NULL;
-    }
-
-    return profile;
-}
-
-/**
- * sgp30_run_profile_by_number() - run a profile by its identifier number
- * @number:     The number that identifies the profile
- *
- * Return:      STATUS_OK on success, an error code otherwise
- */
-static int16_t sgp30_run_profile_by_number(uint16_t number) {
-    const struct sgp_profile *profile;
-    int16_t ret;
-
-    profile = sgp30_get_profile_by_number(number);
-    if (profile == NULL)
-        return STATUS_FAIL;
-
-    ret = sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, profile->command);
+    ret = sgp30_get_feature_set_version(&fs_version, &product_type);
     if (ret != STATUS_OK)
         return ret;
 
-    sensirion_sleep_usec(profile->duration_us);
-
-    if (profile->num_words > 0) {
-        client_data.current_state = MEASURING_PROFILE_STATE;
-        return read_measurement(profile);
-    }
+    if (fs_version < needed_fs)
+        return SGP30_ERR_INSUFFICIENT_FEATURE_SET;
 
     return STATUS_OK;
-}
-
-/**
- * sgp30_detect_featureset_version() - extracts the featureset and initializes
- *                                   client_data.
- *
- * @featureset:  Pointer to the featureset bits
- *
- * Return:    STATUS_OK on success, STATUS_FAIL otherwise
- */
-static int16_t sgp30_detect_featureset_version(uint16_t *featureset) {
-    int16_t i, j;
-    uint16_t feature_set_version = *featureset;
-    const struct sgp_otp_featureset *sgp30_featureset;
-
-    client_data.info.feature_set_version = feature_set_version;
-    client_data.otp_features = &sgp30_features_unknown;
-    for (i = 0; i < sgp_supported_featuresets.number_of_supported_featuresets;
-         ++i) {
-        sgp30_featureset = sgp_supported_featuresets.featuresets[i];
-        for (j = 0;
-             j < sgp30_featureset->number_of_supported_featureset_versions;
-             ++j) {
-            if (SGP_FS_COMPAT(
-                    feature_set_version,
-                    sgp30_featureset->supported_featureset_versions[j])) {
-                client_data.otp_features = sgp30_featureset;
-                return STATUS_OK;
-            }
-        }
-    }
-    return SGP30_ERR_UNKNOWN_FEATURE_SET;
 }
 
 /**
@@ -251,20 +154,7 @@ int16_t sgp30_measure_test(uint16_t *test_result) {
  * Return:  STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_measure_iaq() {
-    const struct sgp_profile *profile;
-    int16_t ret;
-
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_IAQ_MEASURE);
-    if (profile == NULL)
-        return STATUS_FAIL;
-
-    ret = sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, profile->command);
-    if (ret != STATUS_OK)
-        return ret;
-
-    client_data.current_state = MEASURING_PROFILE_STATE;
-
-    return STATUS_OK;
+    return sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, SGP30_CMD_IAQ_MEASURE);
 }
 
 /**
@@ -279,21 +169,16 @@ int16_t sgp30_measure_iaq() {
  * Return:      STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_read_iaq(uint16_t *tvoc_ppb, uint16_t *co2_eq_ppm) {
-    const struct sgp_profile *profile;
     int16_t ret;
+    uint16_t words[SGP30_CMD_IAQ_MEASURE_WORDS];
 
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_IAQ_MEASURE);
-    if (profile == NULL)
-        return STATUS_FAIL;
+    ret = sensirion_i2c_read_words(SGP30_I2C_ADDRESS, words,
+                                   SGP30_CMD_IAQ_MEASURE_WORDS);
 
-    ret = read_measurement(profile);
-    if (ret != STATUS_OK)
-        return ret;
+    *tvoc_ppb = words[1];
+    *co2_eq_ppm = words[0];
 
-    *tvoc_ppb = client_data.buffer.words[0];
-    *co2_eq_ppm = client_data.buffer.words[1];
-
-    return STATUS_OK;
+    return ret;
 }
 
 /**
@@ -310,14 +195,13 @@ int16_t sgp30_measure_iaq_blocking_read(uint16_t *tvoc_ppb,
                                         uint16_t *co2_eq_ppm) {
     int16_t ret;
 
-    ret = sgp30_run_profile_by_number(PROFILE_NUMBER_IAQ_MEASURE);
+    ret = sgp30_measure_iaq();
     if (ret != STATUS_OK)
         return ret;
 
-    *tvoc_ppb = client_data.buffer.words[0];
-    *co2_eq_ppm = client_data.buffer.words[1];
+    sensirion_sleep_usec(SGP30_CMD_IAQ_MEASURE_DURATION_US);
 
-    return STATUS_OK;
+    return sgp30_read_iaq(tvoc_ppb, co2_eq_ppm);
 }
 
 /**
@@ -411,14 +295,13 @@ int16_t sgp30_measure_raw_blocking_read(uint16_t *ethanol_raw_signal,
                                         uint16_t *h2_raw_signal) {
     int16_t ret;
 
-    ret = sgp30_run_profile_by_number(PROFILE_NUMBER_MEASURE_RAW_SIGNALS);
+    ret = sgp30_measure_raw();
     if (ret != STATUS_OK)
         return ret;
 
-    *ethanol_raw_signal = client_data.buffer.words[0];
-    *h2_raw_signal = client_data.buffer.words[1];
+    sensirion_sleep_usec(SGP30_CMD_RAW_MEASURE_DURATION_US);
 
-    return STATUS_OK;
+    return sgp30_read_raw(ethanol_raw_signal, h2_raw_signal);
 }
 
 /**
@@ -430,20 +313,7 @@ int16_t sgp30_measure_raw_blocking_read(uint16_t *ethanol_raw_signal,
  * Return:  STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_measure_raw() {
-    const struct sgp_profile *profile;
-    int16_t ret;
-
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_MEASURE_RAW_SIGNALS);
-    if (profile == NULL)
-        return STATUS_FAIL;
-
-    ret = sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, profile->command);
-    if (ret != STATUS_OK)
-        return ret;
-
-    client_data.current_state = MEASURING_PROFILE_STATE;
-
-    return STATUS_OK;
+    return sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, SGP30_CMD_RAW_MEASURE);
 }
 
 /**
@@ -457,21 +327,16 @@ int16_t sgp30_measure_raw() {
  * Return:      STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_read_raw(uint16_t *ethanol_raw_signal, uint16_t *h2_raw_signal) {
-    const struct sgp_profile *profile;
     int16_t ret;
+    uint16_t words[SGP30_CMD_RAW_MEASURE_WORDS];
 
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_MEASURE_RAW_SIGNALS);
-    if (profile == NULL)
-        return STATUS_FAIL;
+    ret = sensirion_i2c_read_words(SGP30_I2C_ADDRESS, words,
+                                   SGP30_CMD_RAW_MEASURE_WORDS);
 
-    ret = read_measurement(profile);
-    if (ret != STATUS_OK)
-        return ret;
+    *ethanol_raw_signal = words[1];
+    *h2_raw_signal = words[0];
 
-    *ethanol_raw_signal = client_data.buffer.words[0];
-    *h2_raw_signal = client_data.buffer.words[1];
-
-    return STATUS_OK;
+    return ret;
 }
 
 /**
@@ -492,17 +357,28 @@ int16_t sgp30_read_raw(uint16_t *ethanol_raw_signal, uint16_t *h2_raw_signal) {
  * Return:      STATUS_OK on success, else STATUS_FAIL
  */
 int16_t sgp30_get_iaq_baseline(uint32_t *baseline) {
-    int16_t ret = sgp30_run_profile_by_number(PROFILE_NUMBER_IAQ_GET_BASELINE);
+    int16_t ret;
+    uint16_t words[SGP30_CMD_GET_IAQ_BASELINE_WORDS];
+
+    ret =
+        sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, SGP30_CMD_GET_IAQ_BASELINE);
+
     if (ret != STATUS_OK)
         return ret;
 
-    *baseline = (((uint32_t)client_data.buffer.words[0]) << 16) |
-                (uint32_t)client_data.buffer.words[1];
+    sensirion_sleep_usec(SGP30_CMD_GET_IAQ_BASELINE_DURATION_US);
 
-    if (!SGP30_VALID_IAQ_BASELINE(*baseline))
-        return STATUS_FAIL;
+    ret = sensirion_i2c_read_words(SGP30_I2C_ADDRESS, words,
+                                   SGP30_CMD_GET_IAQ_BASELINE_WORDS);
 
-    return STATUS_OK;
+    if (ret != STATUS_OK)
+        return ret;
+
+    *baseline = ((uint32_t)words[1] << 16) | ((uint32_t)words[0]);
+
+    if (*baseline)
+        return STATUS_OK;
+    return STATUS_FAIL;
 }
 
 /**
@@ -516,20 +392,20 @@ int16_t sgp30_get_iaq_baseline(uint32_t *baseline) {
  * Return:      STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_set_iaq_baseline(uint32_t baseline) {
-    const struct sgp_profile *profile;
-    uint16_t words[SENSIRION_NUM_WORDS(baseline)] = {
-        (uint16_t)((baseline & 0xffff0000) >> 16),
-        (uint16_t)(baseline & 0x0000ffff)};
+    int16_t ret;
+    uint16_t words[2] = {(uint16_t)((baseline & 0xffff0000) >> 16),
+                         (uint16_t)(baseline & 0x0000ffff)};
 
-    if (!SGP30_VALID_IAQ_BASELINE(baseline))
+    if (!baseline)
         return STATUS_FAIL;
 
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_IAQ_SET_BASELINE);
-    if (profile == NULL)
-        return STATUS_FAIL;
+    ret = sensirion_i2c_write_cmd_with_args(SGP30_I2C_ADDRESS,
+                                            SGP30_CMD_SET_IAQ_BASELINE, words,
+                                            SENSIRION_NUM_WORDS(words));
 
-    return sensirion_i2c_write_cmd_with_args(
-        SGP30_I2C_ADDRESS, profile->command, words, SENSIRION_NUM_WORDS(words));
+    sensirion_sleep_usec(SGP30_CMD_SET_IAQ_BASELINE_DURATION_US);
+
+    return ret;
 }
 
 /**
@@ -550,14 +426,22 @@ int16_t sgp30_set_iaq_baseline(uint32_t baseline) {
 int16_t sgp30_get_tvoc_inceptive_baseline(uint16_t *tvoc_inceptive_baseline) {
     int16_t ret;
 
-    ret = sgp30_run_profile_by_number(
-        PROFILE_NUMBER_IAQ_GET_TVOC_INCEPTIVE_BASELINE);
+    ret = sgp30_check_featureset(0x21);
+
     if (ret != STATUS_OK)
         return ret;
 
-    *tvoc_inceptive_baseline = client_data.buffer.words[0];
+    ret = sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS,
+                                  SGP30_CMD_GET_TVOC_INCEPTIVE_BASELINE);
 
-    return STATUS_OK;
+    if (ret != STATUS_OK)
+        return ret;
+
+    sensirion_sleep_usec(SGP30_CMD_GET_TVOC_INCEPTIVE_BASELINE_DURATION_US);
+
+    return sensirion_i2c_read_words(
+        SGP30_I2C_ADDRESS, tvoc_inceptive_baseline,
+        SGP30_CMD_GET_TVOC_INCEPTIVE_BASELINE_WORDS);
 }
 
 /**
@@ -570,18 +454,23 @@ int16_t sgp30_get_tvoc_inceptive_baseline(uint16_t *tvoc_inceptive_baseline) {
  * Return:      STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_set_tvoc_baseline(uint16_t tvoc_baseline) {
-    const struct sgp_profile *profile;
+    int16_t ret;
 
-    if (!SGP30_VALID_IAQ_BASELINE(tvoc_baseline))
+    ret = sgp30_check_featureset(0x21);
+
+    if (ret != STATUS_OK)
+        return ret;
+
+    if (!tvoc_baseline)
         return STATUS_FAIL;
 
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_IAQ_SET_TVOC_BASELINE);
-    if (profile == NULL)
-        return STATUS_FAIL;
-
-    return sensirion_i2c_write_cmd_with_args(
-        SGP30_I2C_ADDRESS, profile->command, &tvoc_baseline,
+    ret = sensirion_i2c_write_cmd_with_args(
+        SGP30_I2C_ADDRESS, SGP30_CMD_SET_TVOC_BASELINE, &tvoc_baseline,
         SENSIRION_NUM_WORDS(tvoc_baseline));
+
+    sensirion_sleep_usec(SGP30_CMD_SET_TVOC_BASELINE_DURATION_US);
+
+    return ret;
 }
 
 /**
@@ -596,15 +485,8 @@ int16_t sgp30_set_tvoc_baseline(uint16_t tvoc_baseline) {
  * Return:      STATUS_OK on success, an error code otherwise
  */
 int16_t sgp30_set_absolute_humidity(uint32_t absolute_humidity) {
-    const struct sgp_profile *profile;
+    int16_t ret;
     uint16_t ah_scaled;
-
-    if (!SGP_REQUIRE_FS(client_data.info.feature_set_version, 1, 0))
-        return STATUS_FAIL; /* feature unavailable */
-
-    profile = sgp30_get_profile_by_number(PROFILE_NUMBER_SET_ABSOLUTE_HUMIDITY);
-    if (profile == NULL)
-        return STATUS_FAIL;
 
     if (absolute_humidity > 256000)
         return STATUS_FAIL;
@@ -612,9 +494,13 @@ int16_t sgp30_set_absolute_humidity(uint32_t absolute_humidity) {
     /* ah_scaled = (absolute_humidity / 1000) * 256 */
     ah_scaled = (uint16_t)((absolute_humidity * 16777) >> 16);
 
-    return sensirion_i2c_write_cmd_with_args(SGP30_I2C_ADDRESS,
-                                             profile->command, &ah_scaled,
-                                             SENSIRION_NUM_WORDS(ah_scaled));
+    ret = sensirion_i2c_write_cmd_with_args(
+        SGP30_I2C_ADDRESS, SGP30_CMD_SET_ABSOLUTE_HUMIDITY, &ah_scaled,
+        SENSIRION_NUM_WORDS(ah_scaled));
+
+    sensirion_sleep_usec(SGP30_CMD_SET_ABSOLUTE_HUMIDITY_DURATION_US);
+
+    return ret;
 }
 
 /**
@@ -645,9 +531,20 @@ uint8_t sgp30_get_configured_address() {
  */
 int16_t sgp30_get_feature_set_version(uint16_t *feature_set_version,
                                       uint8_t *product_type) {
-    *feature_set_version = client_data.info.feature_set_version & 0x00FF;
-    *product_type =
-        (uint8_t)((client_data.info.feature_set_version & 0xF000) >> 12);
+    int16_t ret;
+    uint16_t words[SGP30_CMD_GET_FEATURESET_WORDS];
+
+    ret = sensirion_i2c_delayed_read_cmd(SGP30_I2C_ADDRESS,
+                                         SGP30_CMD_GET_FEATURESET,
+                                         SGP30_CMD_GET_FEATURESET_DURATION_US,
+                                         words, SGP30_CMD_GET_FEATURESET_WORDS);
+
+    if (ret != STATUS_OK)
+        return ret;
+
+    *feature_set_version = words[0] & 0x00FF;
+    *product_type = (uint8_t)((words[0] & 0xF000) >> 12);
+
     return STATUS_OK;
 }
 
@@ -659,7 +556,20 @@ int16_t sgp30_get_feature_set_version(uint16_t *feature_set_version,
  * Return:  STATUS_OK on success
  */
 int16_t sgp30_get_serial_id(uint64_t *serial_id) {
-    *serial_id = client_data.info.serial_id;
+    int16_t ret;
+    uint16_t words[SGP30_CMD_GET_SERIAL_ID_WORDS];
+
+    ret = sensirion_i2c_delayed_read_cmd(SGP30_I2C_ADDRESS,
+                                         SGP30_CMD_GET_SERIAL_ID,
+                                         SGP30_CMD_GET_SERIAL_ID_DURATION_US,
+                                         words, SGP30_CMD_GET_SERIAL_ID_WORDS);
+
+    if (ret != STATUS_OK)
+        return ret;
+
+    *serial_id = (((uint64_t)words[0]) << 32) | (((uint64_t)words[1]) << 16) |
+                 (((uint64_t)words[2]) << 0);
+
     return STATUS_OK;
 }
 
@@ -669,7 +579,10 @@ int16_t sgp30_get_serial_id(uint64_t *serial_id) {
  * Return:  STATUS_OK on success.
  */
 int16_t sgp30_iaq_init() {
-    return sgp30_run_profile_by_number(PROFILE_NUMBER_IAQ_INIT);
+    int16_t ret =
+        sensirion_i2c_write_cmd(SGP30_I2C_ADDRESS, SGP30_CMD_IAQ_INIT);
+    sensirion_sleep_usec(SGP30_CMD_IAQ_INIT_DURATION_US);
+    return ret;
 }
 
 /**
@@ -678,40 +591,15 @@ int16_t sgp30_iaq_init() {
  * This call aleady initializes the IAQ baselines (sgp30_iaq_init())
  *
  * Return:  STATUS_OK on success,
- *          SGP30_ERR_UNKNOWN_FEATURE_SET if the sensor's feature set
- *                                        is unknown or outdated,
+ *          SGP30_ERR_INSUFFICIENT_FEATURE_SET if the sensor's feature set
+ *                                             is unknown or outdated,
  *          An error code otherwise
  */
 int16_t sgp30_probe() {
-    int16_t err;
-    uint64_t *serial_buf = &client_data.buffer.u64_value;
+    int16_t ret = sgp30_check_featureset(0x20);
 
-    *serial_buf = 0;
-    client_data.current_state = WAIT_STATE;
-
-    /* try to read the serial ID */
-    err = sensirion_i2c_delayed_read_cmd(
-        SGP30_I2C_ADDRESS, SGP30_CMD_GET_SERIAL_ID,
-        SGP30_CMD_GET_SERIAL_ID_DURATION_US, client_data.buffer.words,
-        SGP30_CMD_GET_SERIAL_ID_WORDS);
-    if (err != STATUS_OK)
-        return err;
-
-    SENSIRION_WORDS_TO_BYTES(client_data.buffer.words,
-                             SGP30_CMD_GET_SERIAL_ID_WORDS);
-    client_data.info.serial_id = be64_to_cpu(*serial_buf) >> 16;
-
-    /* read the featureset version */
-    err = sensirion_i2c_delayed_read_cmd(
-        SGP30_I2C_ADDRESS, SGP30_CMD_GET_FEATURESET,
-        SGP30_CMD_GET_FEATURESET_DURATION_US, client_data.buffer.words,
-        SGP30_CMD_GET_FEATURESET_WORDS);
-    if (err != STATUS_OK)
-        return err;
-
-    err = sgp30_detect_featureset_version(client_data.buffer.words);
-    if (err != STATUS_OK)
-        return err;
+    if (ret != STATUS_OK)
+        return ret;
 
     return sgp30_iaq_init();
 }